// frontend/src/core/api/generated/model/productionItemView.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Kızılkan Şantiye API
 * OpenAPI spec version: 1.0
 */
import type { CrewRef } from './crewRef'
import type { ProductionItemViewStatus } from './productionItemViewStatus'

export interface ProductionItemView {
  id: string
  trade: string
  /** @nullable */
  title?: string | null
  name: string
  crew?: CrewRef
  totalQuantity: number
  unit: string
  doneQuantity: number
  remainingQuantity: number
  percent: number
  todayQuantity: number
  status: ProductionItemViewStatus
  /** @nullable */
  startDate?: string | null
  /** @nullable */
  plannedEnd?: string | null
  /** @nullable */
  note?: string | null
  /** @nullable */
  lastEntryAt?: string | null
  createdAt: string
}